End of a long night, will try and finish device tracking tomorrow

+++ app/src/main/java/au/edu/unimelb/eldercare/user/User.java
@@ -86,7 +86,7 @@ public class User {
         return latitude;
     }
 
-    public void setLatitude(Integer latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
@@ -94,7 +94,7 @@ public class User {
         return longitude;
     }
 
-    public void setLongitude(Integer longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
